feat(apollo-vertex): ai chat bar chart tool
