+++ apps/apollo-vertex/registry.json
@@ -333,6 +333,7 @@
         "@uipath/tooltip"
       ],
       "files": [
+        { "path": "lib/asserts/assert.ts", "type": "registry:lib" },
         {
           "path": "registry/ai-chat/adapters/agenthub/types.ts",
           "type": "registry:lib",
@@ -469,6 +470,11 @@
           "target": "components/ui/ai-chat/charts/distribution-chart-card.tsx"
         },
         {
+          "path": "registry/ai-chat/charts/bar-chart-card.tsx",
+          "type": "registry:ui",
+          "target": "components/ui/ai-chat/charts/bar-chart-card.tsx"
+        },
+        {
           "path": "registry/ai-chat/charts/line-chart-card.tsx",
           "type": "registry:ui",
           "target": "components/ui/ai-chat/charts/line-chart-card.tsx"
@@ -494,19 +500,29 @@
           "target": "components/ui/ai-chat/tools/choices/index.ts"
         },
         {
-          "path": "registry/ai-chat/tools/data-fabric/shared.ts",
+          "path": "registry/ai-chat/tools/data-fabric/util/entities.ts",
           "type": "registry:lib",
-          "target": "components/ui/ai-chat/tools/data-fabric/shared.ts"
+          "target": "components/ui/ai-chat/tools/data-fabric/util/entities.ts"
         },
         {
-          "path": "registry/ai-chat/tools/data-fabric/binned.ts",
+          "path": "registry/ai-chat/tools/data-fabric/util/filters.ts",
           "type": "registry:lib",
-          "target": "components/ui/ai-chat/tools/data-fabric/binned.ts"
+          "target": "components/ui/ai-chat/tools/data-fabric/util/filters.ts"
         },
         {
-          "path": "registry/ai-chat/tools/data-fabric/resolver-result.ts",
+          "path": "registry/ai-chat/tools/data-fabric/util/joins.ts",
           "type": "registry:lib",
-          "target": "components/ui/ai-chat/tools/data-fabric/resolver-result.ts"
+          "target": "components/ui/ai-chat/tools/data-fabric/util/joins.ts"
+        },
+        {
+          "path": "registry/ai-chat/tools/data-fabric/util/chart-helpers.ts",
+          "type": "registry:lib",
+          "target": "components/ui/ai-chat/tools/data-fabric/util/chart-helpers.ts"
+        },
+        {
+          "path": "registry/ai-chat/tools/data-fabric/util/resolver-result.ts",
+          "type": "registry:lib",
+          "target": "components/ui/ai-chat/tools/data-fabric/util/resolver-result.ts"
         },
         {
           "path": "registry/ai-chat/tools/data-fabric/index.ts",
@@ -517,6 +533,11 @@
           "path": "registry/ai-chat/tools/data-fabric-table/data-fabric-table-tool.tsx",
           "type": "registry:ui",
           "target": "components/ui/ai-chat/tools/data-fabric-table/data-fabric-table-tool.tsx"
+        },
+        {
+          "path": "registry/ai-chat/tools/data-fabric-table/table-data-model.ts",
+          "type": "registry:lib",
+          "target": "components/ui/ai-chat/tools/data-fabric-table/table-data-model.ts"
         },
         {
           "path": "registry/ai-chat/tools/data-fabric-table/index.ts",
@@ -532,6 +553,16 @@
           "path": "registry/ai-chat/tools/data-fabric-distribution/index.ts",
           "type": "registry:lib",
           "target": "components/ui/ai-chat/tools/data-fabric-distribution/index.ts"
+        },
+        {
+          "path": "registry/ai-chat/tools/data-fabric-bar/data-fabric-bar-tool.tsx",
+          "type": "registry:ui",
+          "target": "components/ui/ai-chat/tools/data-fabric-bar/data-fabric-bar-tool.tsx"
+        },
+        {
+          "path": "registry/ai-chat/tools/data-fabric-bar/index.ts",
+          "type": "registry:lib",
+          "target": "components/ui/ai-chat/tools/data-fabric-bar/index.ts"
         },
         {
           "path": "registry/ai-chat/tools/data-fabric-line/data-fabric-line-tool.tsx",
