+++ apps/apollo-vertex/registry/ai-chat/tools/data-fabric-bar/index.ts
@@ -1,0 +1,4 @@
+export {
+  createDataFabricBarTool,
+  dataFabricBarClient,
+} from "./data-fabric-bar-tool";
