+++ apps/apollo-vertex/registry/ai-chat/tools/data-fabric/index.ts
@@ -1,12 +1,12 @@
+export { createDataFabricBarTool } from "../data-fabric-bar";
 export { createDataFabricDistributionTool } from "../data-fabric-distribution";
 export { createDataFabricKpiTool } from "../data-fabric-kpi";
 export { createDataFabricLineTool } from "../data-fabric-line";
 export { createDataFabricMultiLineTool } from "../data-fabric-multi-line";
 export { createDataFabricTableTool } from "../data-fabric-table";
 export {
-  buildTableDataModel,
   type DataFabricToolContext,
   type Entity,
   type EntityField,
   generateEntityFieldsDocs,
-} from "./shared";
+} from "./util/entities";
