+++ apps/apollo-vertex/registry/ai-chat/tools/data-fabric/util/resolver-result.ts
